Added Player.cs
This is the Player class where the player's state is stored and where all the command handling is done.  The command handling code is quite messy, lots of nested ifs.  I know that there is a better way to accomplish what I've done and make it look neater, but I was pressed for time so I went with the rather 'mindless' approach and did a lot of copying and pasting.