Added Room.cs
This file contains the Room class which stores the room data.  In hindsight, I could have gotten rid of this class altogether because the 'rooms' and boundaries are all handled in the Player class.  The rooms are technically aisles of a convenience store, so all the navigation is done on an xy plane where every other column is inaccessible because it is the aisle divider.